Improve user lookup by barcode
Not all ILS' log users in with barcode and password, in some modes of
operation.  Here we pull the barcode off the retrieved user data rather
than assuming what we were handed in the "barcode" field is actually
that.

We also store the email address of the user, if the ILS provides it.

Signed-off-by: Mike Rylander <mrylander@gmail.com>