Updated Koha connector
This updates the Koha connector to use various Koha API endpoints where possible.

In particular, it uses the legacy /svc endpoint to create bibs and items. It
uses the Koha REST API for item retrieval and placing holds, but uses the REST
API extensions from https://github.com/NatLibFi/koha-plugin-rest-di to implement
patron authentication, as the Koha REST API currently has no endpoint for
verifying the validity of the credentials other than the API user itself.

It defaults to SIP2 for checkouts and checkins. Resolving Koha bug 23336
would allow the REST API to be used.

Assuming that a single main user is used to access the API, the following
connector parameters are generally needed:

* Connector svc API host (for Koha) - this is new, and should be
  set to the staff hostname
* Connecter svc API user (for Koha) - also new
* Connector svc API password (for Koha) - also new
* Default Connector Host - should be set to the OPAC username
* Default Connector User - this should be set to the OAuth client
  ID of the chosen staff user accessing the API
* Default Connector Password - this should be set to the OAuth
  secret of the staff user accessing the API
* SIP2 Hostname
* SIP2 institution
* SIP2 port
* SIP2 login username
* SIP2 login password

Installation steps need for the Koha-side:

* Install and set up the https://github.com/NatLibFi/koha-plugin-rest-di plugin
* Enable the RESTOAuth2ClientCredentials system preference
* Create a staff user with the following permissions:
  - circulate
  - catalogue
  - borrowers
  - reserveforothers
  - editcatalogue
* Create API keys for that staff user

The connector requires that the Koha version be at least 21.05.

Signed-off-by: Galen Charlton <gmc@equinoxOLI.org>