Improve logic for hiding "Place ILL Request"
Previous to this commit, the logic that hides the "Place ILL Request" links
and buttons was based almost entirely on whether there were any copies that
have a true holdable field. Now we test the shelving location and status
as well, and also check the deleted flags where applicable.

The links/buttons are not hidden if the user is logged in and has the
PLACE_UNFILLABLE_HOLDS permission.

Signed-off-by: Mike Rylander <mrylander@gmail.com>