Allow lender staff to process ILLs for borrowers
Fulfillment depends on both lender and borrower staff to process
materials in the ILL Management interfaces, but there are times when it
is not practical or possible for the borrower to do so.  This commit
adds the ability for lender-side staff to process ILLs for the borrower
in cases where the item is returned to the lender, but the borrower
could not or did not fully process it.

Both the Item Status and Single Scan interfaces are enhanced, with three
new action buttons (including the offered "next action" button in the
Single Scan interface) presented for three slightly different
situations:

  * Perform full Borrower Round-trip: Fulfillment considers the item to
    be in transit to to the borrower, but the borrower has used the
    material and sent it back to the lender without any processing in
    fulfillment.
  * Perform ILL for Borrower: Fulfillment sees that the item is on the
    borrower's ILL (holds) shelf for a patron, but the borrower has not
    recorded a circulation to the patron, and simply sent the item back
    with no processing in Fulfillment beyond the transit receive action.
  * Complete ILL for Borrower: Fulfillment sees an open ILL Transaction
    at the borrower, which the borrower did not complete before sending
    the item back to the lender.

The same underlying mechanism is used for all three situations, but the
incomplete ILL Transaction context is supplied to the lender staff so
that they can inform the borrower of the steps missed, if necessary.

Signed-off-by: Mike Rylander <mrylander@gmail.com>